// packages/redux/src/authentication/types/subArea.types.ts
export type SubAreaType =
  | 'LOGIN'
  | 'LOGOUT'
  | 'PASSWORD_CHANGE'
  | 'PASSWORD_RECOVER'
  | 'PASSWORD_RESET'
  | 'REGISTER'
  | 'VALIDATE_EMAIL'
  | 'REFRESH_EMAIL_TOKEN'
  | 'DELETE_USER_TOKEN'
  | 'DELETE_USER_IMPERSONATION'
  | 'CREATE_USER_IMPERSONATION'
  | 'CREATE_USER_TOKEN'
  | 'CREATE_CLIENT_CREDENTIALS_TOKEN'
  | 'REFRESH_USER_TOKEN';